Massive restructuring of state
Modeled the state much closer to the example in the redux tutorial,
instead of nesting state we normalize games and seasons into their
own top-level items. This will allow us to keep per-game statistics
and implement caching in future commits, as well as elect to fetch or
not fetch games depending on if we have them on the client.