New API module to centralize interacting with the API endpoints
Includes hashing of a defined SECRET, the url, and the current
unix date so the server can verify the call is from our app.
Obviously the SECRET can be found by decompiling for anyone that
wants it but this should be enough to stop naive brute force
attempts on the server at least.