Improved UX for keyboard handling
1) When we have a view with the keyboard, dismiss the keyboard
when the user taps on the screen
2) Use the KeyboardSpacer third party package to automatically slide
up the inputs when the keyboard gains focus and return them when it
loses focus. This involved some extra manipulation since we use flex,
so the invislble KeyboardSpacer contributes to the space-between
justifycontent layout - we take it out of the flex layout using
position: absolute when we don't need it to be there.