Upgrade react-native and dependencies
Wow, this was not fun! I tried to use the git upgrade but ended up
with tons of nasty diffs in my xcode project that I had to resolve
by hand and then there were some breaking changes to the way the
xcode project was built!? After like 3 attempts I decided to create
a new react-native project by hand and merge over my code changes,
excluding bugsnag, which has a new method for the newest version of
react native.

The app boots but haven't done much QA beyond that.