Handle 'and' answers
Require that both be a levenshtein or perfect match - still allow for
a perfect match on the whole phrase of a levenstein match on the whole
phrase.

I cranked up the allowable levenshtein distance here because it was
saying that 'Spain and Algeria' was a lev match for 'Spain and Andorra'
which is too lenient.

Fixes: #9